Move geographic validator methods and tests
